lncli: Add graphical output of routing table
LIGHT-131, LIGHT-140, LIGHT-138
`lncli showroutingtable` may output routing table as image.
Use graphviz for graph rendering.
Add explicit version dependency for tools. Add error checking.

+++ fundingmanager.go
@@ -584,7 +584,7 @@ func (f *fundingManager) handleFundingSignComplete(fmsg *fundingSignCompleteMsg)
 			// so this new channel can be utilized during path
 			// finding.
 			chanInfo := openChan.StateSnapshot()
-			capacity := float64(chanInfo.Capacity)
+			capacity := int64(chanInfo.Capacity)
 			fmsg.peer.server.routingMgr.OpenChannel(
 				graph.NewID(chanInfo.RemoteID),
 				graph.NewEdgeID(fundingPoint.String()),
@@ -654,7 +654,7 @@ func (f *fundingManager) handleFundingOpen(fmsg *fundingOpenMsg) {
 		resCtx.reservation.FundingOutpoint, fmsg.peer.id)
 
 	// Notify the L3 routing manager of the newly active channel link.
-	capacity := float64(resCtx.reservation.OurContribution().FundingAmount +
+	capacity := int64(resCtx.reservation.OurContribution().FundingAmount +
 		resCtx.reservation.TheirContribution().FundingAmount)
 	fmsg.peer.server.routingMgr.OpenChannel(
 		graph.NewID([32]byte(fmsg.peer.lightningID)),
